Removed mistake, timeout in update service and added await for intern…

+++ install/update.sh
@@ -13,35 +13,46 @@
 
 set -o xtrace
 
-echo "Attempting to resolve project..."
-if cd /home/seard/client/duktig ; then
-    echo "Client exists => trying to pull latest version from Github repository..."
-    if sudo -H -u seard bash -c 'git pull -f' ; then
-        echo "Latest version pulled successfully"
+echo "Awaiting internet"
+if for i in {1..20}; do ping -c1 www.google.com &> /dev/null && break; done ; then
+    echo "Internet!!!"
+
+    echo "Attempting to resolve project..."
+    if cd /home/seard/client/duktig ; then
+        echo "Client exists => trying to pull latest version from Github repository..."
+        if sudo -H -u seard bash -c 'git pull -f' ; then
+            echo "Latest version pulled successfully"
+        else
+            echo "Failed when pulling => deleting folder and cloning new repository..."
+            git status
+            cd /home/
+            sudo rm -rf /home/seard/client
+            sudo -H -u seard bash -c 'mkdir /home/seard/client'
+            sudo -H -u seard bash -c 'git clone https://github.com/seard/duktig.git /home/seard/client/'
+        fi
     else
-        echo "Failed when pulling => deleting folder and cloning new repository..."
-        git status
+        echo "Client does not exist => creating it"
         cd /home/
         sudo rm -rf /home/seard/client
         sudo -H -u seard bash -c 'mkdir /home/seard/client'
         sudo -H -u seard bash -c 'git clone https://github.com/seard/duktig.git /home/seard/client/'
     fi
+
+    if cd /home/seard/client/install/ ; then
+        echo "Setting up services..."
+        sudo ./install.sh
+
+        cd /home/seard/client/duktig
+
+        echo "Running npm install..."
+        sudo -H -u seard bash -c 'timeout 10m npm install'
+    else
+        echo "Broken"
+    fi
+
 else
-    echo "Client does not exist => creating it"
-    cd /home/
-    sudo rm -rf /home/seard/client
-    sudo -H -u seard bash -c 'mkdir /home/seard/client'
-    sudo -H -u seard bash -c 'git clone https://github.com/seard/duktig.git /home/seard/client/'
+    echo "No internet :("
 fi
-
-echo "Setting up services..."
-cd /home/seard/client/install/
-sudo ./install.sh
-
-cd /home/seard/client/duktig
-
-echo "Running npm install..."
-sudo -H -u seard bash -c 'timeout 10m npm install'
 
 #date && echo git checkout master
 #date && echo git checkout
